+++ source/ci_source/providers/index.ts
@@ -6,8 +6,9 @@ import {FakeCI} from "./Fake"
 import {Surf} from "./Surf"
 import {DockerCloud} from "./DockerCloud"
 import {Codeship} from "./Codeship"
+import {Drone} from "./Drone"
 
-const providers: Array<any> = [Travis, Circle, Semaphore, Jenkins, FakeCI, Surf, DockerCloud, Codeship]
+const providers: Array<any> = [Travis, Circle, Semaphore, Jenkins, FakeCI, Surf, DockerCloud, Codeship, Drone]
 export {
   providers
 };