+++ changelog.md
@@ -1,6 +1,7 @@
 ### master
 
 //  Add your own contribution below
+* Add support for [Drone](http://readme.drone.io) - gabro
 
 ### 0.12.0